requirements.txt
fastapi==0.70.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
sqlalchemy==1.4.48
uvicorn==0.15.0
pydantic
psycopg2==2.9.1
python-dotenv
